# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "YADMB",
    "description": "The web api for interacting with the music bot"
  },
  "tags": [
    {
      "name": "favorites",
      "description": "Manage and see the user's favorites"
    },
    {
      "name": "queue",
      "description": "Manage the server's queue"
    },
    {
      "name": "song",
      "description": "Play/pause the current song"
    }
  ],
  "paths": {
    "/favorites": {
      "get": {
        "tags": [
          "favorites"
        ],
        "summary": "Get favorites",
        "description": "Returns all the favorite for the given user",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "getFavorites",
        "responses": {
          "200": {
            "description": "All the favorites of the user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/Favorite1"
                      },
                      {
                        "$ref": "#/components/schemas/Favorite2"
                      },
                      {
                        "$ref": "#/components/schemas/Favorite3"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token not valid"
          }
        }
      },
      "post": {
        "tags": [
          "favorites"
        ],
        "summary": "Add a favorite",
        "description": "Adds a favorite to the database",
        "parameters": [
          {
            "name": "name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 0,
              "maxLength": 100
            },
            "description": "The name of the favorite, max 100 characters"
          },
          {
            "name": "link",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 0,
              "maxLength": 200
            },
            "description": "The link of the song, max 200 characters"
          },
          {
            "name": "folder",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 0,
              "maxLength": 100
            },
            "description": "The name of the folder"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "addFavorite",
        "responses": {
          "200": {
            "description": "Favorite added successfully"
          },
          "401": {
            "description": "Token not valid"
          },
          "500": {
            "description": "Name duplicated for the given user"
          }
        }
      },
      "delete": {
        "tags": [
          "favorites"
        ],
        "summary": "Remove a favorite",
        "description": "Removes a favorite from the database",
        "parameters": [
          {
            "name": "name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 0,
              "maxLength": 100
            },
            "description": "The name of the favorite"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "removeFavorite",
        "responses": {
          "200": {
            "description": "Favorite removed successfully"
          },
          "401": {
            "description": "Token not valid"
          },
          "500": {
            "description": "Favorite doesn't exist"
          }
        }
      }
    },
    "/queue/{guild}": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get the entire queue for a given guild",
        "description": "Returns the entire queue of a guild",
        "parameters": [
          {
            "name": "guild",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The guild id"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "getQueue",
        "responses": {
          "200": {
            "description": "Elements of the queue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/Element1"
                      },
                      {
                        "$ref": "#/components/schemas/Element2"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token not valid or it doesn't have access to the given guild"
          }
        }
      },
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Add a song to the queue",
        "description": "Adds the given song to the queue",
        "parameters": [
          {
            "name": "guild",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The guild id"
          },
          {
            "name": "song",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "A link to a song or a query to search on youtube"
          },
          {
            "name": "playlist",
            "in": "header",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": "false"
            },
            "description": "Is the song a playlist?"
          },
          {
            "name": "shuffle",
            "in": "header",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": "false"
            },
            "description": "Should we shuffle the given playlist?"
          },
          {
            "name": "loop",
            "in": "header",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": "false"
            },
            "description": "Should the song be looped?"
          },
          {
            "name": "priority",
            "in": "header",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": "false"
            },
            "description": "Should we add the song as the first in the queue?"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "addToQueue",
        "responses": {
          "200": {
            "description": "Song added successfully"
          },
          "401": {
            "description": "Token not valid or it doesn't have access to the given guild"
          },
          "403": {
            "description": "User not in voice channel of the guild, or dj mode is active and user is not a dj"
          },
          "406": {
            "description": "Playlist supplied, but flag was not raised"
          }
        }
      },
      "delete": {
        "tags": [
          "queue"
        ],
        "summary": "Skip the currently playing song",
        "description": "Skip the currently playing song",
        "parameters": [
          {
            "name": "guild",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The guild id"
          },
          {
            "name": "clean",
            "in": "header",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": "false"
            },
            "description": "Should we clean the entire queue?"
          }
        ],
        "operationId": "skip",
        "responses": {
          "200": {
            "description": "Song skipped successfully"
          },
          "401": {
            "description": "Token not valid or it doesn't have access to the given guild"
          },
          "403": {
            "description": "User not in voice channel of the guild"
          },
          "406": {
            "description": "The bot is not playing anything or the queue is empty"
          }
        }
      }
    },
    "/song/pause/{guild}": {
      "get": {
        "tags": [
          "song"
        ],
        "summary": "Pause the currently playing song",
        "description": "Pause the currently playing song",
        "parameters": [
          {
            "name": "guild",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The guild id"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "pause",
        "responses": {
          "200": {
            "description": "Song paused successfully"
          },
          "401": {
            "description": "Token not valid or it doesn't have access to the given guild"
          },
          "406": {
            "description": "Song is already paused or the queue is empty"
          }
        }
      }
    },
    "/song/resume/{guild}": {
      "get": {
        "tags": [
          "song"
        ],
        "summary": "Resume the currently playing song",
        "description": "Resume the currently playing song",
        "parameters": [
          {
            "name": "guild",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The guild id"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "resume",
        "responses": {
          "200": {
            "description": "Song resumed successfully"
          },
          "401": {
            "description": "Token not valid or it doesn't have access to the given guild"
          },
          "406": {
            "description": "Song is already playing  or the queue is empty"
          }
        }
      }
    },
    "/song/toggle/{guild}": {
      "get": {
        "tags": [
          "song"
        ],
        "summary": "Play/pauses the currently playing song",
        "description": "Play/pauses the currently playing song",
        "parameters": [
          {
            "name": "guild",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The guild id"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token for authentication"
          }
        ],
        "operationId": "toggle",
        "responses": {
          "200": {
            "description": "Song paused or resumed successfully"
          },
          "401": {
            "description": "Token not valid or it doesn't have access to the given guild"
          },
          "406": {
            "description": "The queue is empty"
          },
          "500": {
            "description": "Couldn't toggle song"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Element1": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "KygYQJEi3SQ-youtube"
          },
          "title": {
            "type": "string",
            "example": "Charlie - Faccia da Pirla"
          },
          "duration": {
            "type": "string",
            "example": "04:19"
          },
          "link": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=KygYQJEi3SQ"
          },
          "user": {
            "type": "string",
            "example": "thetipo01"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://i.ytimg.com/vi/KygYQJEi3SQ/maxresdefault.jpg"
          },
          "loop": {
            "type": "boolean",
            "example": false
          },
          "frames": {
            "type": "integer",
            "example": 10248
          }
        },
        "required": [
          "id",
          "title",
          "duration",
          "link",
          "user",
          "thumbnail",
          "loop"
        ]
      },
      "Element2": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PvuYSybooLg-youtube"
          },
          "title": {
            "type": "string",
            "example": "Lights, Camera, Action! - Studiopolis Zone Act 1"
          },
          "duration": {
            "type": "string",
            "example": "03:08"
          },
          "link": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=PvuYSybooLg"
          },
          "user": {
            "type": "string",
            "example": "thetipo01"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://i.ytimg.com/vi/PvuYSybooLg/maxresdefault.jpg"
          },
          "loop": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "title",
          "duration",
          "link",
          "user",
          "thumbnail",
          "loop"
        ]
      },
      "Favorite1": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Very nice song"
          },
          "link": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=PvuYSybooLg"
          },
          "folder": {
            "type": "string",
            "example": "Sonic OST"
          }
        },
        "required": [
          "name",
          "link"
        ]
      },
      "Favorite2": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "The Top"
          },
          "link": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=6wftUIvoHAQ"
          },
          "folder": {
            "type": "string",
            "example": "Eurobeat"
          }
        },
        "required": [
          "name",
          "link"
        ]
      },
      "Favorite3": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Blue Jeans"
          },
          "link": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=4MKAf6YX_7M"
          },
          "folder": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "name",
          "link"
        ]
      }
    }
  }
}